+++ demos/manual/src/main/java/is/codion/manual/store/test/StoreLoadTest.java
@@ -48,7 +48,7 @@ public class StoreLoadTest {
 			EntityConnectionProvider connectionProvider =
 							RemoteEntityConnectionProvider.builder()
 											.user(user)
-											.domainType(Store.DOMAIN)
+											.domain(Store.DOMAIN)
 											.build();
 
 			return new StoreApplicationModel(connectionProvider);