+++ framework/db-http/src/test/java/is/codion/framework/db/http/HttpEntityConnectionProviderTest.java
@@ -33,7 +33,7 @@ public class HttpEntityConnectionProviderTest {
 		EntityConnectionProvider.CLIENT_CONNECTION_TYPE.set(EntityConnectionProvider.CONNECTION_TYPE_HTTP);
 		try {
 			EntityConnectionProvider connectionProvider = EntityConnectionProvider.builder()
-							.domainType(TestDomain.DOMAIN)
+							.domain(TestDomain.DOMAIN)
 							.clientType("test")
 							.user(User.parse("scott:tiger"))
 							.build();